pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["Hwannni <cmhcms1115@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
pyjwt = "^2.9.0"
django-environ = "^0.11.2"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
